add blueprint to the report
